fix not export authOptions from route.ts
