+++ frontend/src/app/api/auth/[...nextauth]/route.ts
@@ -1,72 +1,5 @@
-import type { LoginAsStudentResponse } from '@/lib/backend/types/auth-type'
-import NextAuth, { type NextAuthOptions } from 'next-auth'
-import CredentialsProvider from 'next-auth/providers/credentials'
-
-export const authOptions: NextAuthOptions = {
-  secret: process.env.NEXTAUTH_SECRET,
-  providers: [
-    CredentialsProvider({
-      name: 'Credentials',
-      credentials: {
-        email: { label: 'メールアドレス', type: 'email' },
-      },
-      async authorize(credentials) {
-        try {
-          const response = await fetch(
-            `${process.env.BACKEND_URL}/login-as-student`,
-            {
-              method: 'POST',
-              headers: {
-                'Content-Type': 'application/json',
-              },
-              body: JSON.stringify({
-                Email: credentials?.email,
-              }),
-            },
-          )
-
-          if (!response.ok) {
-            return null
-          }
-
-          const data: LoginAsStudentResponse = await response.json()
-          return {
-            id: data.StudentID.toString(),
-            token: data.Token,
-            name: data.Name,
-          }
-        } catch (error) {
-          console.error('Login error:', error)
-          return null
-        }
-      },
-    }),
-  ],
-  callbacks: {
-    async jwt({ token, user }) {
-      if (user) {
-        token.accessToken = user.token
-        token.id = user.id
-      }
-      return token
-    },
-    async session({ session, token }) {
-      if (token) {
-        session.accessToken = token.accessToken
-      }
-      return {
-        ...session,
-        user: {
-          ...session.user,
-          id: token.id,
-        },
-      }
-    },
-  },
-  pages: {
-    signIn: '/login',
-  },
-}
+import NextAuth from 'next-auth'
+import { authOptions } from './options'
 
 const handler = NextAuth(authOptions)
