+++ frontend/src/app/student/page.tsx
@@ -1,4 +1,4 @@
-import { authOptions } from '@/app/api/auth/[...nextauth]/route'
+import { authOptions } from '@/app/api/auth/[...nextauth]/options'
 import { getStudentInfo } from '@/lib/backend/student'
 import { getServerSession } from 'next-auth/next'
 import Link from 'next/link'
