+++ frontend/src/app/student/tasks/page.tsx
@@ -1,4 +1,4 @@
-import { authOptions } from '@/app/api/auth/[...nextauth]/route'
+import { authOptions } from '@/app/api/auth/[...nextauth]/options'
 import { getTasksByStudent } from '@/lib/backend/task'
 import { getServerSession } from 'next-auth'
 import { redirect } from 'next/navigation'
